Optimized the way the plugin send a message to the players
I created a function that manage the way the plugin communicate with the player. Being by dealing with the communication between 2 players, or by sending error messages.
Instead of always using sendMessage, i'm now using a function that take 3 parameters : the target, the color of the message and the message itself.

I didn't save space, but I'm now more flexible.